backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
catboost==1.2
shap==0.44.1
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
pytesseract==0.3.10
opencv-python==4.9.0.80
pdf2image==1.16.3
Pillow==10.2.0
werkzeug==3.0.1
gunicorn==21.2.0
